Separate classes into headers.
