+++ .vscode/settings.json
@@ -1,0 +1,19 @@
+{
+    "files.associations": {
+        "__config": "cpp",
+        "__nullptr": "cpp",
+        "cstddef": "cpp",
+        "exception": "cpp",
+        "initializer_list": "cpp",
+        "new": "cpp",
+        "stdexcept": "cpp",
+        "type_traits": "cpp",
+        "typeinfo": "cpp",
+        "iostream": "cpp",
+        "memory": "cpp",
+        "utility": "cpp",
+        "algorithm": "cpp",
+        "vector": "cpp",
+        "cstdlib": "cpp"
+    }
+}
