+++ exercises/t_jubb/ex02_oo_basics/CMakeLists.txt
@@ -18,10 +18,22 @@ include ( CommonSetup )
 
 set ( SRC_FILES
   src/main.cpp
+  src/Shape.cpp
+  src/Square.cpp
+  src/Circle.cpp
+  src/Rectangle.cpp
+  src/IsoscelesTriangle.cpp
+  src/ShapeSorter.cpp
   # add any additional source files here
 )
 
 set ( HDR_FILES
+  src/Shape.h
+  src/Square.h
+  src/Circle.h
+  src/Rectangle.h
+  src/IsoscelesTriangle.h
+  src/ShapeSorter.h
   # add any additional header files here
 )
