+++ exercises/t_jubb/ex02_oo_basics/src/Circle.h
@@ -1,0 +1,20 @@
+#ifndef HEADER_CIRCLE
+#define HEADER_CIRCLE
+
+#include "Shape.h"
+
+/*
+Circle Class
+*/
+class Circle : public Shape {
+
+public:
+	Circle(double r);
+	double area();
+	double perimeter();
+
+private:
+	double radius;
+};
+
+#endif
