+++ exercises/t_jubb/ex02_oo_basics/src/Square.h
@@ -1,0 +1,22 @@
+#ifndef HEADER_SQUARE
+#define HEADER_SQUARE
+
+#include "Shape.h"
+
+/*
+Square class
+*/
+class Square : public Shape {
+
+public:
+	Square(double h);
+	double area() ;
+	double perimeter();
+
+private:
+	double height;
+};
+
+
+
+#endif
